+++ src/ruliology_forge/__init__.py
@@ -1,5 +1,6 @@
 """Ruliology Forge: computational resilience tools for rule-based systems."""
 
+from .benchmarks import BenchmarkScenario, BenchmarkSuite, benchmark_rules, rank_benchmark_rows, standard_suite
 from .eca import ECARule, initial_state, simulate_eca
 from .experiments import ExperimentConfig, PerturbationResult, run_perturbation_experiment, scan_rules
 from .metrics import (
@@ -18,10 +19,13 @@ from .perturb import perturb_state
 from .plotting import plot_divergence, plot_trajectory
 
 __all__ = [
+    "BenchmarkScenario",
+    "BenchmarkSuite",
     "ECARule",
     "ExperimentConfig",
     "PerturbationResult",
     "ResilienceSummary",
+    "benchmark_rules",
     "divergence_auc",
     "final_restoration",
     "hamming_divergence",
@@ -29,6 +33,7 @@ __all__ = [
     "perturb_state",
     "plot_divergence",
     "plot_trajectory",
+    "rank_benchmark_rows",
     "recovery_time",
     "restoration_coefficient",
     "run_perturbation_experiment",
@@ -36,6 +41,7 @@ __all__ = [
     "scar_metrics",
     "shift_tolerant_divergence",
     "simulate_eca",
+    "standard_suite",
     "summarize_resilience",
     "xor_difference",
 ]